separate model utilities
